+++ packages/moment-tz-date-adapter/src/lib/moment-tz-date-adapter.ts
@@ -59,7 +59,9 @@ export class MomentTZDateAdapter extends DateAdapter implements IDateAdapter<mom
     return MomentTZDateAdapter.fromJSON(datetime.toJSON());
   }
 
-  readonly date: moment.Moment;
+  get date() {
+    return this._date.clone();
+  }
   readonly timezone: string | null;
   readonly duration: number | undefined;
   readonly generators: unknown[] = [];
@@ -67,12 +69,13 @@ export class MomentTZDateAdapter extends DateAdapter implements IDateAdapter<mom
   protected readonly [MOMENT_TZ_DATE_ADAPTER_ID] = true;
 
   private _end: moment.Moment | undefined;
+  private _date: moment.Moment;
 
   constructor(date: moment.Moment, options: { duration?: number } = {}) {
     super(undefined);
 
-    this.date = date.clone();
-    this.timezone = this.date.tz() || null;
+    this._date = date.clone();
+    this.timezone = date.tz() || null;
     this.duration = options.duration;
 
     this.assertIsValid();
@@ -93,7 +96,7 @@ export class MomentTZDateAdapter extends DateAdapter implements IDateAdapter<mom
   set(_: 'timezone', value: string | null) {
     if (this.timezone === value) return this;
 
-    const date = this.date.clone();
+    const date = this._date.clone();
 
     if (value === null) {
       // work around for https://github.com/moment/moment-timezone/issues/738
@@ -106,11 +109,11 @@ export class MomentTZDateAdapter extends DateAdapter implements IDateAdapter<mom
   }
 
   valueOf() {
-    return this.date.valueOf();
+    return this._date.valueOf();
   }
 
   toISOString() {
-    return this.date.toISOString();
+    return this._date.toISOString();
   }
 
   toDateTime(): DateTime {
@@ -121,18 +124,18 @@ export class MomentTZDateAdapter extends DateAdapter implements IDateAdapter<mom
     return {
       timezone: this.timezone,
       duration: this.duration,
-      year: this.date.get('year'),
-      month: this.date.get('month') + 1,
-      day: this.date.get('date'),
-      hour: this.date.get('hour'),
-      minute: this.date.get('minute'),
-      second: this.date.get('second'),
-      millisecond: this.date.get('millisecond'),
+      year: this._date.get('year'),
+      month: this._date.get('month') + 1,
+      day: this._date.get('date'),
+      hour: this._date.get('hour'),
+      minute: this._date.get('minute'),
+      second: this._date.get('second'),
+      millisecond: this._date.get('millisecond'),
     };
   }
 
   assertIsValid() {
-    if (!this.date.isValid()) {
+    if (!this._date.isValid()) {
       throw new InvalidDateAdapterError();
     } else if (this.duration && this.duration <= 0) {
       throw new InvalidDateAdapterError('If provided, duration must be greater than 0.');